# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vanila Meteo Api",
        "description": "Vanila Meteo Api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:99/api",
            "description": "local server"
        },
        {
            "url": "http://staging.example.com",
            "description": "staging server"
        },
        {
            "url": "http://example.com",
            "description": "production server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "description": "Login user and generate token",
                "operationId": "19649594a8d95f07aecffa517ffcd224",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new user",
                "operationId": "895cc6dd7beae96f66b99bcafc6bc341",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout authenticated user",
                "description": "Revokes the current user's authentication token",
                "operationId": "89f11567c9c2915976bb5864b1fadc60",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/user": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get user details",
                "description": "Get the authenticated user's details",
                "operationId": "fda1506d38419f2a3b6f6f4eb0fc379e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User data retrieved"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ciudades": {
            "get": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Get all cities",
                "operationId": "8e9234b4172f03a70a78b3a9b9f07d93",
                "responses": {
                    "200": {
                        "description": "List of all cities"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Create new city",
                "operationId": "189bbe1246ce262ffb41bf6e00cce784",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "latitud",
                                    "longitud"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "City created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/ciudades/forecast": {
            "get": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Generate forecast for all cities",
                "operationId": "1ff0f4f847bd12f71ef637f38c4a15c0",
                "responses": {
                    "200": {
                        "description": "Weather data stored successfully"
                    },
                    "500": {
                        "description": "Error generating forecast"
                    }
                }
            }
        },
        "/ciudades/{id}": {
            "get": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Get city by ID",
                "operationId": "55b99645b31d26addc2529251ac395b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City details"
                    },
                    "404": {
                        "description": "City not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Update city",
                "operationId": "c247226e19b59fb411c97972144c57ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "latitud",
                                    "longitud"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "City updated successfully"
                    },
                    "404": {
                        "description": "City not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Delete city",
                "operationId": "bff7f9e24d20d94a9d04d6b517d6a3fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City deleted successfully"
                    },
                    "404": {
                        "description": "City not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pronostico/guardar": {
            "post": {
                "tags": [
                    "Pronóstico"
                ],
                "summary": "Save weather forecast",
                "operationId": "dfe542ca8415e3356d3161fb8b61e72e",
                "parameters": [
                    {
                        "name": "ciudad_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Forecast saved successfully"
                    }
                }
            }
        },
        "/pronostico/{ciudad_nombre}/{fecha_inicio}/{fecha_fin}": {
            "get": {
                "tags": [
                    "Pronóstico"
                ],
                "summary": "Get forecast by city name and date range",
                "operationId": "ed153da394f2e60d184c43cc9111e73b",
                "parameters": [
                    {
                        "name": "ciudad_nombre",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fecha_inicio",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_fin",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Forecast data retrieved"
                    },
                    "404": {
                        "description": "City not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pronostico/ciudad/{ciudad_nombre}": {
            "get": {
                "tags": [
                    "Pronóstico"
                ],
                "summary": "Get forecast by city name",
                "operationId": "c0bceea0cd5eea1b407a6f65125d488f",
                "parameters": [
                    {
                        "name": "ciudad_nombre",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Forecast data retrieved"
                    },
                    "404": {
                        "description": "City not found"
                    },
                    "500": {
                        "description": "API Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Ciudades",
            "description": "Ciudades"
        },
        {
            "name": "Pronóstico",
            "description": "Pronóstico"
        }
    ]
}